# frontend/packages/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "需脉枢纽 API",
    "description": "需脉枢纽 · B2B 代工制造供需智能语义匹配平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户注册（厂商/买家）",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthToken_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户登录",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthToken_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/send-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "发送验证码",
        "operationId": "send_code_api_v1_auth_send_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "当前用户信息",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserOut_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/vendors/{vendor_id}/audit": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "厂商能力档案审核",
        "operationId": "audit_vendor_api_v1_admin_vendors__vendor_id__audit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuditResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/vendors": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "厂商列表（按审核状态筛选，分页）",
        "operationId": "list_vendors_api_v1_admin_vendors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "audit_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "pending/passed/rejected",
              "title": "Audit Status"
            },
            "description": "pending/passed/rejected"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_VendorListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "数据概览（四个统计卡片）",
        "operationId": "stats_api_v1_admin_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AdminStatsResponse_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/requests": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "需求与匹配查看（分页）",
        "operationId": "list_requests_api_v1_admin_requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AdminRequestListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/buyers": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "买家列表（搜索/状态筛选，分页）",
        "operationId": "list_buyers_api_v1_admin_buyers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "手机号/邮箱关键词",
              "title": "Keyword"
            },
            "description": "手机号/邮箱关键词"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "active/disabled",
              "title": "Status"
            },
            "description": "active/disabled"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BuyerListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "事件日志（审计，按动作筛选，分页）",
        "operationId": "list_logs_api_v1_admin_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "动作类型",
              "title": "Action"
            },
            "description": "动作类型"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AdminLogListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminRequestListResponse": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/AdminRequestItem"
            },
            "type": "array",
            "title": "List"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "list",
          "total",
          "page",
          "page_size"
        ],
        "title": "AdminRequestListResponse"
      },
      "AdminStatsResponse": {
        "properties": {
          "total_users": {
            "type": "integer",
            "title": "Total Users"
          },
          "total_requests": {
            "type": "integer",
            "title": "Total Requests"
          },
          "total_vendors": {
            "type": "integer",
            "title": "Total Vendors"
          },
          "total_matches": {
            "type": "integer",
            "title": "Total Matches"
          }
        },
        "type": "object",
        "required": [
          "total_users",
          "total_requests",
          "total_vendors",
          "total_matches"
        ],
        "title": "AdminStatsResponse",
        "description": "数据概览（产品 3.1 AC1 四个统计卡片）。"
      },
      "ApiResponse_AdminLogListResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminLogListResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[AdminLogListResponse]"
      },
      "ApiResponse_AdminRequestListResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminRequestListResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[AdminRequestListResponse]"
      },
      "ApiResponse_AdminStatsResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminStatsResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[AdminStatsResponse]"
      },
      "ApiResponse_AuditResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuditResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[AuditResponse]"
      },
      "ApiResponse_AuthToken_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthToken"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[AuthToken]"
      },
      "ApiResponse_BuyerListResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuyerListResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[BuyerListResponse]"
      },
...
      "ApiResponse_UserOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[UserOut]"
      },
      "ApiResponse_VendorListResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VendorListResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[VendorListResponse]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "AuditRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "pass",
              "reject"
            ],
            "title": "Action"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "审核意见（驳回时必填）"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "AuditRequest"
      },
      "AuditResponse": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "audit_status": {
            "type": "string",
            "enum": [
              "passed",
              "rejected"
            ],
            "title": "Audit Status"
          },
          "audited_at": {
            "type": "string",
            "format": "date-time",
            "title": "Audited At"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "audit_status",
          "audited_at"
        ],
        "title": "AuditResponse"
      },
      "AuthToken": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "AuthToken"
      },
...
      "BuyerListResponse": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/BuyerItem"
            },
            "type": "array",
            "title": "List"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "list",
          "total",
          "page",
          "page_size"
        ],
        "title": "BuyerListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "手机号（与 email 至少其一）"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "邮箱"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "enum": [
              "vendor",
              "buyer"
            ],
            "title": "Role",
            "default": "buyer"
          },
          "verify_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Code",
            "description": "短信/邮箱验证码"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "RegisterRequest",
        "description": "自助注册：手机号/邮箱 + 密码 + 角色 + 验证码（厂商/买家）。"
      },
...
      "SendCodeRequest": {
        "properties": {
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "scene": {
            "type": "string",
            "enum": [
              "register",
              "login",
              "reset"
            ],
            "title": "Scene",
            "default": "register"
          }
        },
        "type": "object",
        "title": "SendCodeRequest"
      },
...
      "UserOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role",
          "created_at"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VendorListResponse": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/VendorAuditItem"
            },
            "type": "array",
            "title": "List"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "list",
          "total",
          "page",
          "page_size"
        ],
        "title": "VendorListResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}